Remove second argument to RTCPeerConnection
It looks like constraints are no longer passed into the
`RTCPeerConnection` object, only to `createOffer` and `createAnswer`.
Spec link here:
https://w3c.github.io/webrtc-pc/#rtcpeerconnection-interface

I think we should stop passing the second argument into the
`RTCPeerConnection` constructor and instead treat `constraints` as a
default that is used when the more specific `offerConstraints` or
`answerConstraints` are not specified.