Do not override code property of thrown errors
When external code invoked by this module throws an error, it's passed
to the `makeError` function which adds a `.code` property to the
error.

I've come across a problem where the thrown error already has a `.code`
property, and worse it's been marked read-only so this module then
throws when it tries to overwrite the property, completely obscuring the
original problem.

The change in this PR is to create a new error, set the `.code` property
on it, and where the error was caused by another error being thrown,
keep the wrapped error as a `.cause` property on the error we throw.

It uses the `err-code` module to do so as a one-liner in order to keep
everything concise.

This also fixes a usability  bug this module has, in that because the
errors thrown by this module are all created in the `makeError`
function, the stack trace of the thrown error starts there instead of
where the error actually occurred, where `makeError` was invoked.